Some more instructions

+++ chip8/test/EmulatorTest.java
@@ -7,7 +7,7 @@ public class EmulatorTest {
     @Test
     public void test() {
         Emulator emulator = new Emulator();
-        int[] data = new int[]{0x1F,0xFF};
+        byte[] data = new byte[]{0x1F,(byte) 0xFF};
         emulator.loadProgram(data);
         emulator.next();
         assertEquals(0xFFF,emulator.i);
